requirements.txt
## requirements.txt
## MIT License
## Created Date: 2024-09-01
## Updated Date: 2025-05-03
## Version 1.2

requests
matplotlib
openai
python-dotenv
fastapi
uvicorn
pydantic
apscheduler
selenium
beautifulsoup4
lxml

# tkinter ships with most Python installers. If it is missing, install it
# through your OS/Python distribution rather than pip.
